Moved product components under product folder and renamed slider to carousel
